백엔드 useThinking 사용
