+++ src/nest/ai/gemini/translator/services/gemini-translator.service.ts
@@ -5,40 +5,38 @@ import {
   GoogleGenerativeAIFetchError,
 } from '@google/generative-ai';
 import { Inject, Injectable } from '@nestjs/common';
-
 import { GeminiModel } from '../../../../../ai/gemini/gemini-models';
-import { getDefaultMaxOutputTokenCountByModel } from '../../../../../ai/model';
 import { TranslationResult } from '../../../../../types/translators';
 import { keyRoundRobin } from '../../../../../utils/key-round-robin';
 import { SourceLanguage } from '../../../../../utils/language';
 import { sleep } from '../../../../../utils/sleep';
 import { tagTexts } from '../../../../../utils/string';
-import { AiTranslatorService } from '../../../common/services/ai-translator-service';
-import { AiTokenService } from '../../../common/services/ai-token-service';
 import { ICacheManagerService } from '../../../../cache/cache-manager/services/i-cache-manager-service';
 import { LoggerService } from '../../../../logger/logger.service';
 import { ExampleManagerService } from '../../../../translation/example/services/example-manager.service';
 import { GeminiPromptConverterService } from '../../prompt/services/gemini-prompt-converter.service';
-import { GeminiTokenService } from './gemini-token.service';
+import { AiTokenService } from './gemini-token.service';
 import { GeminiResponseService } from './gemini-response.service';
 import { AiResponseService } from '../../../common/services/ai-response-service';
 import { FilePathInfo } from '@/types/cache';
 import { AiTranslateParam } from '@/nest/ai/common/services/i-ai-translator-service';
+import { deepClone } from '@/utils/deep-clone';
+import { isNullish } from '@/utils/is-nullish';
+import { RateLimiter } from 'limiter';
 
 @Injectable()
-export class GeminiTranslatorService extends AiTranslatorService<GeminiModel, GenerativeModel> {
+export class GeminiTranslatorService {
+  protected rateLimiterMapping: Map<string, RateLimiter> = new Map();
   private readonly MAX_ATTEMPT_COUNT = 3;
   constructor(
     @Inject(ICacheManagerService) protected readonly cacheManagerService: ICacheManagerService,
-    @Inject(GeminiTokenService) private readonly tokenService: AiTokenService<GeminiModel>,
+    private readonly tokenService: AiTokenService,
     @Inject(GeminiResponseService)
     private readonly geminiResponseService: AiResponseService<EnhancedGenerateContentResponse>,
     private readonly logger: LoggerService,
     protected readonly exampleManagerService: ExampleManagerService,
     private readonly promptConverterService: GeminiPromptConverterService
-  ) {
-    super();
-  }
+  ) {}
 
   protected async getDefaultRequestsPerMinute(modelName: GeminiModel): Promise<number> {
     switch (modelName) {
@@ -63,6 +61,7 @@ export class GeminiTranslatorService extends AiTranslatorService<GeminiModel, Ge
       requestsPerMinute,
       apiKey,
       promptPresetContent,
+      useThinking,
     }: AiTranslateParam
   ): Promise<string[]> {
     this.setRateLimiter(modelName, requestsPerMinute);
@@ -81,8 +80,8 @@ export class GeminiTranslatorService extends AiTranslatorService<GeminiModel, Ge
           const remainingTextArray = Array.from(currentRemainingTexts.keys());
           const batchGroups = await this.tokenService.getBatchGroups({
             texts: remainingTextArray,
-            maxOutputTokenCount: maxOutputTokenCount,
-            modelName,
+            maxOutputTokenCount,
+            useThinking,
           });
 
           for (const batchTexts of batchGroups) {
@@ -225,18 +224,16 @@ export class GeminiTranslatorService extends AiTranslatorService<GeminiModel, Ge
   }: {
     modelName: GeminiModel;
     apiKeyIterator: Generator<string>;
-    maxOutputTokenCount?: number;
+    maxOutputTokenCount: number;
   }): Promise<GenerativeModel> {
     const apiKey = apiKeyIterator.next().value;
-
-    const maxOutputTokens = getDefaultMaxOutputTokenCountByModel(modelName, maxOutputTokenCount);
 
     return new GoogleGenerativeAI(apiKey).getGenerativeModel({
       model: modelName,
       safetySettings: [],
       generationConfig: {
         temperature: 0.5,
-        maxOutputTokens,
+        maxOutputTokens: maxOutputTokenCount,
         topP: 0.95,
       },
     });
@@ -255,7 +252,7 @@ export class GeminiTranslatorService extends AiTranslatorService<GeminiModel, Ge
     sourceLanguage: SourceLanguage;
     modelName: GeminiModel;
     apiKeyIterator: Generator<string>;
-    maxOutputTokenCount?: number;
+    maxOutputTokenCount: number;
     fileInfo?: FilePathInfo;
     promptPresetContent?: string;
   }): Promise<{
@@ -311,5 +308,106 @@ export class GeminiTranslatorService extends AiTranslatorService<GeminiModel, Ge
 
   public async getEstimatedTokenCount(texts: string[] | string): Promise<number> {
     return await this.tokenService.getEstimatedTokenCount(texts);
+  }
+
+  protected async setRateLimiter(modelName: string, requestsPerMinute: number): Promise<void> {
+    if (this.rateLimiterMapping.has(modelName)) return;
+
+    this.rateLimiterMapping.set(
+      modelName,
+      new RateLimiter({
+        tokensPerInterval: requestsPerMinute,
+        interval: 'minute',
+      })
+    );
+  }
+
+  protected async getRateLimiter(modelName: string): Promise<RateLimiter> {
+    if (!this.rateLimiterMapping.has(modelName)) await this.setRateLimiter(modelName, 100);
+    return this.rateLimiterMapping.get(modelName)!;
+  }
+
+  protected async applyTranslationCache(sourceTexts: string[]): Promise<{
+    texts: string[];
+    remainingTexts: Map<string, number[]>;
+  }> {
+    const texts = new Array<string>(sourceTexts.length);
+    const remainingTexts = new Map<string, number[]>();
+    const cachedResults = await this.cacheManagerService.getTranslations(sourceTexts);
+
+    sourceTexts.forEach((text, index) => {
+      const { translatedText, isCacheHit } = this.getTranslationFromCachedResult(
+        text,
+        cachedResults
+      );
+
+      if (isCacheHit) {
+        texts[index] = translatedText;
+      } else {
+        const indices = remainingTexts.get(text) || [];
+        indices.push(index);
+        remainingTexts.set(text, indices);
+      }
+    });
+
+    return { texts, remainingTexts };
+  }
+
+  protected getTranslationFromCachedResult(
+    originalText: string,
+    cachedResults: Map<string, string | null>
+  ): { translatedText: string; isCacheHit: boolean } {
+    const lineRemovedText = originalText
+      .replaceAll('\r', '')
+      .replaceAll('\n', '')
+      .replaceAll('\\n', '')
+      .replaceAll('\\r', '');
+    const trimmedText = lineRemovedText.trim();
+    if (trimmedText === '') return { translatedText: originalText, isCacheHit: true };
+
+    const cachedTranslation = cachedResults.get(trimmedText);
+    const translatedText = isNullish(cachedTranslation) ? originalText : cachedTranslation;
+    const isCacheHit = !isNullish(cachedTranslation);
+
+    return { translatedText, isCacheHit };
+  }
+
+  protected async updateTranslationsAndCache({
+    newTranslations,
+    translations,
+    sourceLanguage,
+    modelName,
+    fileInfo,
+  }: {
+    newTranslations: Map<string, TranslationResult>;
+    translations: string[];
+    sourceLanguage: SourceLanguage;
+    modelName: string;
+    fileInfo?: FilePathInfo;
+  }): Promise<string[]> {
+    const translationsToCache = new Map<string, string>();
+    const sourceLines: string[] = [];
+    const resultLines: string[] = [];
+    const copiedTranslations = deepClone(translations);
+
+    for (const [originalText, { text: translatedText, indices }] of newTranslations) {
+      translationsToCache.set(originalText, translatedText);
+
+      indices.forEach((index) => {
+        copiedTranslations[index] = translatedText;
+      });
+
+      sourceLines.push(originalText);
+      resultLines.push(translatedText);
+    }
+
+    // 번역 저장 - 이제 자동으로 이력도 생성됨
+    await this.cacheManagerService.setTranslations(translationsToCache, true, fileInfo, modelName);
+
+    if (sourceLines.length > 0) {
+      this.exampleManagerService.appendCurrentExample(sourceLanguage, sourceLines, resultLines);
+    }
+
+    return copiedTranslations;
   }
 }
